python: In parsing args of write method of CrFile use int for length instead of Py_ssize_t.

+++ src/python/compression_wrapper-py.c
@@ -196,7 +196,7 @@ static PyObject *
 py_write(_CrFileObject *self, PyObject *args)
 {
     char *str;
-    Py_ssize_t len;
+    int len;
     GError *tmp_err = NULL;
 
     if (!PyArg_ParseTuple(args, "s#:set_num_of_pkgs", &str, &len))
